feat: Refactor OrderService and CartItemService to integrate with new data structure

+++ src/main/java/com/hcmut/bkuety/dto/orders/OrderResponseDTO.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -12,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderResponseDTO {
-    private Double total;
+    private BigDecimal total;
     private String paymentMethod;
     private LocalDate orderDate;
     private String address;
